requirements.txt
pandas==0.24.1
keras==2.2.4
numpy==1.16.2
tensorflow-gpu==1.12.0
Pillow==5.4.1